Fixes compile warnings in public headers. (Tested on gcc44, linux)
The issues fall into 3 types:
* Implicit casts (such as treating an int as a float).  These are
changed to explicit static_casts.

* Declaration of function parameter names sometimes shadow internal
members.  This is addressed by adding an underscore (_) suffix to the
function parameter names as needed.

* Functions implemented in public headers have occasional unused
parameters.  This has been fixed by commenting out the variable names in
the declaration, /* unused */.   Or when not appropriate, I've moved the
implementation from the header to the .cpp file. (Which is preferable in
some cases anyways).

Review by: lg

git-svn-id: http://svn.openimageio.org/oiio/branches/RB-0.9@1774 31904243-a364-42ac-a3db-9305396518ba

+++ src/include/fmath.h
@@ -720,7 +720,7 @@ floorfrac (float x, int *xint)
     int i = FloorToInt (x);
 #endif
     *xint = i;
-    return x - i;   // Return the fraction left over
+    return x - static_cast<float>(i);   // Return the fraction left over
 }
 
 
@@ -819,7 +819,7 @@ float_to_rational (float f, unsigned int &num, unsigned int &den)
     } else {
         num = (int)f;
         den = 1;
-        while (fabsf(f-num) > 0.00001 && den < 1000000) {
+        while (fabsf(f-static_cast<float>(num)) > 0.00001f && den < 1000000) {
             den *= 10;
             f *= 10;
             num = (int)f;
